Add files via upload
This dataset (source) consists of data from almost 10,000 borrowers that took loans - with some paid back and others still in progress. It was extracted from lendingclub.com which is an organization that connects borrowers with investors. 
Goal is to extract useful insights and visualize them in the most interesting way possible, and build a model that can predict the probability a user will be able to pay back their loan within a certain period.