Add files via upload
This dataset (https://www.kaggle.com/itssuru/loan-data) consists of data from almost 10,000 borrowers that took loans - with some paid back and others still in progress. It was extracted from lendingclub.com which is an organization that connects borrowers with investors.
I explore what kind of purposes users take loans for, visualise insights and  build a model that can predict the probability a user will be able to pay back their loan within a certain period.